server: Post-novo server cleanup
