+++ package.json
@@ -40,9 +40,9 @@
     "sentry": "scripts/sentry.sh",
     "start": "scripts/start.sh",
     "start:dev": "concurrently 'yarn start:dev:relay' 'yarn start:dev:server'",
-    "start:dev:server": "NODE_ENV=development node --max_old_space_size=3072 -r dotenv/config ./src/index.dev.js",
+    "start:dev:server": "NODE_ENV=development node --max_old_space_size=3072 -r dotenv/config ./src/dev.js",
     "start:dev:relay": "NODE_ENV=development yarn relay --watch",
-    "start:dev:debug": "NODE_ENV=development node --inspect --max_old_space_size=3072 -r dotenv/config ./src/index.dev.js",
+    "start:dev:debug": "NODE_ENV=development node --inspect --max_old_space_size=3072 -r dotenv/config ./src/dev.js",
     "start:prod": "yarn build:assets:legacy && yarn build:assets && yarn build:server && NODE_ENV=production yarn start",
     "start:prod:debug": "NODE_ENV=production node --inspect --max_old_space_size=3072 -r dotenv/config ./server.dist.js",
     "sync-schema": "scripts/sync-schema-pull.sh",
