+++ src/index.js
@@ -1,63 +1,46 @@
-/* eslint-disable no-console */
-// TODO: Does source map support provide additional server side value?
-import "source-map-support/register"
-
-// TODO: Is the regenerator runtime still required?
-import "regenerator-runtime/runtime"
+// @ts-check
 
 // This must come before any other instrumented module.
 // See https://docs.datadoghq.com/tracing/languages/nodejs/ for more info.
 import "./lib/datadog"
+import "dotenv/config"
 
-import { once } from "lodash"
+// Needs to be first, due to sharify side-effects.
+import { initializeMiddleware } from "./middleware"
+
 import express from "express"
-import http from "http"
-import withGracefulShutdown from "http-shutdown"
-import { initialize } from "./common-app"
+import config from "./config"
+import chalk from "chalk"
+import { startServer } from "./lib/startServer"
+import { app as errorHandling } from "lib/middleware/errorHandling"
+import legacyDesktopApp from "./desktop"
+import legacyMobileApp from "./mobile"
+import forceV2 from "./v2/server"
+
+console.log(chalk.green(`\n[Force] NODE_ENV=${config.NODE_ENV}\n`))
 
 const app = express()
-const commonApp = initialize(once(startServer))
-app.use(commonApp)
 
-// Run start server as a callback for now until waiting on the xapp token and
-// cache have a chance to be refactored.
-function startServer() {
-  const message =
-    process.env.NODE_ENV === "development"
-      ? `\n\n  [Force] Booting on port ${process.env.PORT}... \n`
-      : `\n\n  [Force] Started on ${process.env.APP_URL}. \n`
+initializeMiddleware(app)
 
-  const server = withGracefulShutdown(http.createServer(app))
+// Mount v2 Force
+app.use("/", forceV2)
 
-  const stopServer = once(() => {
-    server.shutdown(() => {
-      console.log("Closed existing connections.")
-      process.exit(0)
-    })
-  })
-
-  if (process.env.KEEPALIVE_TIMEOUT_SECONDS) {
-    console.log(
-      "Setting keepAliveTimeout to " +
-        process.env.KEEPALIVE_TIMEOUT_SECONDS +
-        " sec."
-    )
-    server.keepAliveTimeout =
-      Number(process.env.KEEPALIVE_TIMEOUT_SECONDS) * 1000
+// Mount legacy mobile apps
+app.use((req, res, next) => {
+  if (res.locals.sd.IS_MOBILE) {
+    // Mount mobile app
+    legacyMobileApp(req, res, next)
+  } else {
+    next()
   }
+})
 
-  if (process.env.HEADERS_TIMEOUT_SECONDS) {
-    console.log(
-      "Setting headersTimeout to " +
-        process.env.HEADERS_TIMEOUT_SECONDS +
-        " sec."
-    )
-    server.headersTimeout = Number(process.env.HEADERS_TIMEOUT_SECONDS) * 1000
-  }
+// Mount legacy desktop apps
+app.use(legacyDesktopApp)
 
-  server.listen(process.env.PORT, "0.0.0.0", () => console.log(message))
+// Error handling goes last
+app.use(errorHandling)
 
-  process.on("SIGTERM", stopServer)
-}
-
-module.exports = app
+// Boot server
+startServer(app)
